# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elite Care Medical API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/webhook:
    get:
      operationId: verifyWhatsAppWebhook
      tags: [whatsapp]
      summary: Verify the Meta WhatsApp webhook
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook challenge
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Invalid webhook verification
    post:
      operationId: receiveWhatsAppWebhookPayload
      tags: [whatsapp]
      summary: Receive inbound WhatsApp messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                object:
                  type: string
                entry:
                  type: array
                  items:
                    type: object
                    properties:
                      changes:
                        type: array
                        items:
                          type: object
                          properties:
                            value:
                              type: object
                              properties:
                                messages:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      from:
                                        type: string
                                      id:
                                        type: string
                                      type:
                                        type: string
                                      text:
                                        type: object
                                        properties:
                                          body:
                                            type: string
      responses:
        "200":
          description: Webhook accepted
        "401":
          description: Invalid signature
...